j11 partie 2

+++ 2022/j11/src/inputs.rs
@@ -1,7 +1,7 @@
 #[allow(dead_code)]
 pub struct Monkey
 {
-    pub test: usize,
+    pub test: u64,
     pub dest: (usize, usize)
 }
 
@@ -18,19 +18,19 @@ pub fn example() -> Vec<Monkey> {
 #[allow(dead_code)]
 pub fn input() -> Vec<Monkey> {
     vec![
-        Monkey { test: 2, dest: (5, 2) },
+        Monkey { test: 2,  dest: (5, 2) },
         Monkey { test: 13, dest: (4, 3) },
-        Monkey { test: 5, dest: (5, 1) },
-        Monkey { test: 3, dest: (6, 7) },
+        Monkey { test: 5,  dest: (5, 1) },
+        Monkey { test: 3,  dest: (6, 7) },
         Monkey { test: 11, dest: (7, 3) },
         Monkey { test: 17, dest: (4, 1) },
-        Monkey { test: 7, dest: (0, 2) },
+        Monkey { test: 7,  dest: (0, 2) },
         Monkey { test: 19, dest: (6, 0) },
     ]
 }
 
 #[allow(dead_code)]
-pub fn operate_example (monkey_number: usize, item: usize) -> usize {
+pub fn operate_example (monkey_number: u64, item: u64) -> u64 {
     match monkey_number {
         0 => { item * 19 },
         1 => { item + 6 },
@@ -41,7 +41,7 @@ pub fn operate_example (monkey_number: usize, item: usize) -> usize {
 }
 
 #[allow(dead_code)]
-pub fn operate_input (monkey_number: usize, item: usize) -> usize {
+pub fn operate_input (monkey_number: u64, item: u64) -> u64 {
     match monkey_number {
         0 => { item * 13 },
         1 => { item + 7 },
